[18706] US #601: filtrering i joboversigten på sink - Rettelse i JobListCriteriaModelMapper således at denne nu konverterer string værdien for sink id til int.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@87945 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ gui/src/main/java/dk/dbc/dataio/gui/server/modelmappers/criterias/JobListCriteriaModelMapper.java
@@ -45,7 +45,7 @@ public final class JobListCriteriaModelMapper {
      * @param model the job list criteria model
      */
     private static void buildJobListCriteriaWithSinkClause(JobListCriteria jobListCriteria, JobListCriteriaModel model) {
-        jobListCriteria.where(new ListFilter<JobListCriteria.Field>(JobListCriteria.Field.SINK_ID, ListFilter.Op.EQUAL, model.getSinkId()));
+        jobListCriteria.where(new ListFilter<JobListCriteria.Field>(JobListCriteria.Field.SINK_ID, ListFilter.Op.EQUAL, Long.valueOf(model.getSinkId()).intValue()));
     }
 
     /**
